Move wrapper-component-manufacturing functions into a reusable utility
